RavenDB-21860 code cleanup
